requirements.txt
pynput
numpy
pyautogui
opencv-python
pillow